pkg/config: Split impl from options and integrate flag handler including tests
