+++ pkg/config/build.go
@@ -10,22 +10,3 @@ type BuildOptions struct {
 func NewBuildOptions() *BuildOptions {
 	return &BuildOptions{}
 }
-
-// BuildImpl is impl for applyOptions
-type BuildImpl struct {
-	*GlobalImpl
-	*BuildOptions
-}
-
-// NewBuildImpl creates a new BuildImpl
-func NewBuildImpl(g *GlobalImpl, b *BuildOptions) *BuildImpl {
-	return &BuildImpl{
-		GlobalImpl:   g,
-		BuildOptions: b,
-	}
-}
-
-// EmbedValues returns the embed values.
-func (b *BuildImpl) EmbedValues() bool {
-	return b.BuildOptions.EmbedValues
-}
