+++ pkg/config/cache_impl.go
@@ -1,0 +1,15 @@
+package config
+
+// CacheImpl is impl for CacheOptions
+type CacheImpl struct {
+	*GlobalImpl
+	*CacheOptions
+}
+
+// NewCacheImpl creates a new CacheImpl
+func NewCacheImpl(g *GlobalImpl, b *CacheOptions) *CacheImpl {
+	return &CacheImpl{
+		GlobalImpl:   g,
+		CacheOptions: b,
+	}
+}
