+++ pkg/config/show-dag.go
@@ -8,17 +8,3 @@ type ShowDAGOptions struct {
 func NewShowDAGOptions() *ShowDAGOptions {
 	return &ShowDAGOptions{}
 }
-
-// ShowDAGImpl is impl for applyOptions
-type ShowDAGImpl struct {
-	*GlobalImpl
-	*ShowDAGOptions
-}
-
-// NewShowDAGImpl creates a new ShowDAGImpl
-func NewShowDAGImpl(g *GlobalImpl, b *ShowDAGOptions) *ShowDAGImpl {
-	return &ShowDAGImpl{
-		GlobalImpl:     g,
-		ShowDAGOptions: b,
-	}
-}
