+++ pkg/config/show-dag_impl.go
@@ -1,0 +1,15 @@
+package config
+
+// ShowDAGImpl is impl for applyOptions
+type ShowDAGImpl struct {
+	*GlobalImpl
+	*ShowDAGOptions
+}
+
+// NewShowDAGImpl creates a new ShowDAGImpl
+func NewShowDAGImpl(g *GlobalImpl, b *ShowDAGOptions) *ShowDAGImpl {
+	return &ShowDAGImpl{
+		GlobalImpl:     g,
+		ShowDAGOptions: b,
+	}
+}
