+++ pkg/config/template.go
@@ -1,11 +1,8 @@
 package config
 
 import (
-	"fmt"
-	"os"
-	"strings"
-
 	"github.com/helmfile/helmfile/pkg/common"
+	"github.com/helmfile/helmfile/pkg/flags"
 )
 
 // TemplateOptions is the options for the build command
@@ -51,132 +48,28 @@ type TemplateOptions struct {
 	ShowOnly []string
 }
 
-// NewTemplateOptions creates a new Apply
+// NewTemplateOptions creates a new TemplateOption
 func NewTemplateOptions() *TemplateOptions {
-	newOptions := &TemplateOptions{
-		SkipCRDsFlag:    common.NewBoolFlag(false),
-		IncludeCRDsFlag: common.NewBoolFlag(false),
-	}
-	return newOptions
+	options := &TemplateOptions{}
+	options.Initialize()
+	return options
 }
 
-// TemplateImpl is impl for applyOptions
-type TemplateImpl struct {
-	*GlobalImpl
-	*TemplateOptions
+func (o *TemplateOptions) Initialize() {
+	flags.EnsureBoolFlag(&o.IncludeCRDsFlag, false)
+	flags.EnsureBoolFlag(&o.SkipCRDsFlag, false) // not exposed as cli flag but needed for ShouldIncludeCRDs() until skip-crds is removed
 }
 
-// NewTemplateImpl creates a new TemplateImpl
-func NewTemplateImpl(g *GlobalImpl, t *TemplateOptions) *TemplateImpl {
-	return &TemplateImpl{
-		GlobalImpl:      g,
-		TemplateOptions: t,
-	}
-}
-
-// Concurrency returns the concurrency
-func (t *TemplateImpl) Concurrency() int {
-	return t.TemplateOptions.Concurrency
-}
-
-// SkipCRDs returns the skip crds
-func (t *TemplateImpl) SkipCRDs() bool {
-	return t.TemplateOptions.SkipCRDsFlag.Value()
-}
-
-// IncludeCRDs returns the include crds
-func (t *TemplateImpl) IncludeCRDs() bool {
-	return t.TemplateOptions.IncludeCRDsFlag.Value()
-}
-
-// ShouldIncludeCRDs returns whether to include crds
-func (t *TemplateImpl) ShouldIncludeCRDs() bool {
-	includeCRDsExplicit := t.IncludeCRDsFlag.WasExplicitlySet() && t.IncludeCRDsFlag.Value()
-	skipCRDsExplicit := t.SkipCRDsFlag.WasExplicitlySet() && !t.SkipCRDsFlag.Value()
-
-	return includeCRDsExplicit || skipCRDsExplicit
-}
-
-// NoHooks returns the no hooks
-func (t *TemplateImpl) NoHooks() bool {
-	return t.TemplateOptions.NoHooks
-}
-
-// IncludeNeeds returns the include needs
-func (t *TemplateImpl) IncludeNeeds() bool {
-	return t.TemplateOptions.IncludeNeeds || t.IncludeTransitiveNeeds()
-}
-
-// IncludeTransitiveNeeds returns the include transitive needs
-func (t *TemplateImpl) IncludeTransitiveNeeds() bool {
-	return t.TemplateOptions.IncludeTransitiveNeeds
-}
-
-// OutputDir returns the output dir
-func (t *TemplateImpl) OutputDir() string {
-	return strings.TrimRight(t.TemplateOptions.OutputDir, fmt.Sprintf("%c", os.PathSeparator))
-}
-
-// OutputDirTemplate returns the output dir template
-func (t *TemplateImpl) OutputDirTemplate() string {
-	return t.TemplateOptions.OutputDirTemplate
-}
-
-// Set returns the Set
-func (t *TemplateImpl) Set() []string {
-	return t.TemplateOptions.Set
-}
-
-// SkipCleanup returns the skip cleanup
-func (t *TemplateImpl) SkipCleanup() bool {
-	return t.TemplateOptions.SkipCleanup
-}
-
-// SkipNeeds returns the skip needs
-func (t *TemplateImpl) SkipNeeds() bool {
-	if !t.IncludeNeeds() {
-		return t.TemplateOptions.SkipNeeds
+func (o *TemplateOptions) HandleFlag(name string, value interface{}, changed bool) bool {
+	switch name {
+	case "include-crds":
+		if changed {
+			if boolVal, ok := value.(*bool); ok {
+				o.IncludeCRDsFlag.Set(*boolVal)
+			}
+		}
+		return true
 	}
 
 	return false
-}
-
-// SkipTests returns the skip tests
-func (t *TemplateImpl) SkipTests() bool {
-	return t.TemplateOptions.SkipTests
-}
-
-// Validate returns the validate
-func (t *TemplateImpl) Validate() bool {
-	return t.TemplateOptions.Validate
-}
-
-// Values returns the values
-func (t *TemplateImpl) Values() []string {
-	return t.TemplateOptions.Values
-}
-
-// PostRenderer returns the PostRenderer.
-func (t *TemplateImpl) PostRenderer() string {
-	return t.TemplateOptions.PostRenderer
-}
-
-// PostRendererArgs returns the PostRendererArgs.
-func (t *TemplateImpl) PostRendererArgs() []string {
-	return t.TemplateOptions.PostRendererArgs
-}
-
-// SkipSchemaValidation returns the SkipSchemaValidation.
-func (t *TemplateImpl) SkipSchemaValidation() bool {
-	return t.TemplateOptions.SkipSchemaValidation
-}
-
-// KubeVersion returns the the KubeVersion.
-func (t *TemplateImpl) KubeVersion() string {
-	return t.TemplateOptions.KubeVersion
-}
-
-// ShowOnly returns the ShowOnly.
-func (t *TemplateImpl) ShowOnly() []string {
-	return t.TemplateOptions.ShowOnly
 }
